UAS-112 added endpoint to register a sale transaction

+++ src/main/kotlin/ucb/accounting/backend/dao/SaleTransactionDetail.kt
@@ -21,7 +21,7 @@ class SaleTransactionDetail {
     var subaccountId: Long = 0
 
     @Column(name = "quantity")
-    var quantity: Long = 0
+    var quantity: Int = 0
 
     @Column(name = "unit_price_bs")
     var unitPriceBs: BigDecimal = BigDecimal.ZERO
